Do not delete digest file when packaging

+++ WalletWasabi.Packager/Program.cs
@@ -422,7 +422,7 @@ namespace WalletWasabi.Packager
 
 				foreach (var file in torFolder.EnumerateFiles())
 				{
-					if (!file.Name.Contains("data", StringComparison.OrdinalIgnoreCase) && !file.Name.Contains(toNotRemove, StringComparison.OrdinalIgnoreCase))
+					if (!file.Name.Contains("data", StringComparison.OrdinalIgnoreCase) && !file.Name.Contains("digest", StringComparison.OrdinalIgnoreCase) && !file.Name.Contains(toNotRemove, StringComparison.OrdinalIgnoreCase))
 					{
 						File.Delete(file.FullName);
 					}
